backend/flask_server_2/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
Werkzeug==3.0.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
torch==2.2.2
torchvision==0.17.2
vit-pytorch==1.7.12
einops==0.8.0
